CRUD operations for issues
Updated urls.py with urls for issues and forms.py with a form for issues. Created html templates for one issue preview, form for creating a new issue and updating an existing issue and updated the html template for project preview with the list of issues. Added views for CRUD operations for issues.