CRUD operations for milestone
Updated urls.py with urls for milestone and forms.py with form for
milestone. Created html templates for one milestone preview, form for
creating a new milestone and updating an existing milestone. Updated the
html template for project preview with list of milestones. Added views
for CRUD operations for milestone.