CRUD operations for labels
Edited the issue preview template to show labels. Created forms for creating a new label and choosing one of the existing labels in the project. Created templates for adding a new label to the issue and choosing one of the existing labels in the project for the issue. Added views and urls for CRUD operations for labels. Changed the relationship between issue and label from ManyToOne to ManyToMany.